refactor : 컴포넌트 정리 및 페이지 정리
