--- gabojago_fe/src/components/compo/Tokyo.vue
@@ -1,7 +1,0 @@
-<template>
-  <div></div>
-</template>
-
-<script setup></script>
-
-<style scoped></style>
