+++ gabojago_fe/src/components/unused/Text.vue
