+++ gabojago_fe/src/components/used/TopbarWithIcon-deleteFull.vue
