--- gabojago_fe/src/pages/Modal_test.vue
@@ -1,6 +1,0 @@
-<template></template>
-
-<script setup>
-import { ref } from 'vue';
-import Modal from './components/Modal.vue';
-</script>
