--- gabojago_fe/src/pages/Siwan_tour.vue
@@ -1,28 +1,0 @@
-<template>
-  <div></div>
-  <div>
-    <HistoryListItem
-      list="관광"
-      number="-5,109"
-      number2="-580"
-      img="/src/assets/관광.png"
-    />
-
-    <HistoryListItem
-      list="현금"
-      number="+1,000,000"
-      number2="+11,392.77"
-      img="/src/assets/현금.png"
-    />
-  </div>
-</template>
-
-<script setup>
-// import TopbarXSign test from "@/components/TopbarXSign.vue";
-
-// import Tokyo from "@/components/Tokyo.vue";
-import HistoryListItem from '@/components/compo/HistoryListItem.vue';
-// import DeleteButton-siwan from "@/componets/DeleteButton-siwan.vue";
-</script>
-
-<style scoped></style>
