+++ gabojago_fe/src/pages/Trip/Trip_AddCalendar.vue
@@ -17,9 +17,9 @@
 </template>
 
 <script setup>
-import Topbar from '@/components/compo/Topbar.vue';
-import Calendar from '@/components/compo/Calendar.vue';
-import CtaBar from '@/components/compo/CtaBar.vue';
+import Topbar from '@/components/used/Topbar.vue';
+import Calendar from '@/components/used/Calendar.vue';
+import CtaBar from '@/components/used/CtaBar.vue';
 import { useRouter, useRoute } from 'vue-router';
 import { onMounted, ref } from 'vue';
 import { useAddTripStore } from '@/stores/tripStore';
