+++ gabojago_fe/src/pages/Trip/Trip_AddDescription.vue
@@ -18,9 +18,9 @@
 </template>
 
 <script setup>
-import Topbar from '@/components/compo/Topbar.vue';
-import TextInput2 from '@/components/compo/TextInput2.vue';
-import CtaBar from '@/components/compo/CtaBar.vue';
+import Topbar from '@/components/used/Topbar.vue';
+import TextInput2 from '@/components/used/TextInput2.vue';
+import CtaBar from '@/components/used/CtaBar.vue';
 import { useRouter, useRoute } from 'vue-router';
 import { onMounted, ref } from 'vue';
 import { useAddTripStore } from '@/stores/tripStore';
