+++ gabojago_fe/src/pages/Trip/Trip_AddPayment.vue
@@ -88,12 +88,12 @@
 
 <script setup>
 import { ref, onMounted } from 'vue';
-import Topbar from '@/components/compo/Topbar.vue';
-import CtaBar from '@/components/compo/CtaBar.vue';
-import TopSelect from '@/components/compo/TopSelect.vue';
+import Topbar from '@/components/used/Topbar.vue';
+import CtaBar from '@/components/used/CtaBar.vue';
+import TopSelect from '@/components/used/TopSelect.vue';
 import { useRouter, useRoute } from 'vue-router';
 import { getCurrency, postTransaction } from '@/api/transaction';
-import Modal from '@/components/compo/Modal.vue';
+import Modal from '@/components/used/Modal.vue';
 import { useAddTripStore } from '@/stores/tripStore';
 
 const router = useRouter();
@@ -163,7 +163,7 @@ const registerExpense = async () => {
     console.error('Failed to register expense:', error);
   }
 
-  router.push('/dashbaord');
+  router.push({ path: '/dashboard' });
 };
 
 async function convertCurrency() {
